BLANK! as _ in DELIMIT, SPACED, UNSPACED, "" opts-out
Before the change that made NULL the basic currency for opting out,
BLANK!s were frequently generated by routines like ANY and ALL.  That
along with the idea that a variable might be empty made it seem
desirable that a PRINT or forming operation (like SPACED, UNSPACED, or
DELIMIT) would treat blanks as invisible.

With NULL taking the place of true opting out, the tides turned to
where TO TEXT! of a BLANK! was really a BLANK!.  This allowed for some
usages of blanks to be using them for their "underscore-ness" instead
of for their "blankness".

One detail made this difficult, in that SPACED could not be a
specialization of DELIMIT with the delimiter as NULL.  This is because
SPECIALIZE does not permit specialization with null arguments.  Hence,
an empty string "" was changed to be treated as a no-op w.r.t. spacing.