PARSE returns null/BAR! by default, allow THROW of ERROR!
PARSE is not strictly a LOGIC!-returning function, and can return
anything with RETURN.  Hence, it is useful to make its failure map
to null instead of false, so it can be used e.g. like:

    value: <not-found> unless parse data [
        some 'foo return (<some-foos-found>)
    ]

This changes the failure to be a null.  The default return of success
is a BAR! value, which is what is typically used for "synthesized"
truthy values...as returning #[true] would give the incorrect
impression that the failure case would be #[false].  Hence it cues
users to go to either a DID PARSE or NOT PARSE (if they were interested
in a LOGIC!) or a TRY PARSE (if they were going to return a value or
fail otherwise, and want to put that value into a variable).

Also, speculatively makes PARSE able to take a BLANK! value in and then
return a NULL out...which makes it join in the "blank in, null out"
protocol that is followed by many non-mutating routines.

!!! Since PARSE can sometimes mutate, this is a little dodgy, so it
should be reviewed if it turns out to seem misleading in some cases...
depends on how useful people find it in the meantime.

Given the recent tendency toward fewer restrictions "for the safety of
the programmer", this takes away the restriction that THROW can't
throw a value of type ERROR!.  This was added when there was widespread
belief that THROW MAKE ERROR! was a good way to invoke the error (the
actual previous correct way was DO MAKE ERROR!).  Ren-C has FAIL, and
it is less of an issue...so allowing a full-spectrum throw seems the
better path at this point.