Fix bug related to stale f->gotten pointer
This fixes what was apparently a longstanding bug, which would happen
in the circumstance of:

    do [(lots of code) some-var-code-relocates-in-memory]

Basically, the WORD! is fetched before running the GROUP! in a
lookahead step, and stored in a frame member called `f->gotten`.  That
pointer is into an object's varlist, which in the general case can be
expanded and thus invalidate the pointer.  The idea is to reuse the
cache where possible, but since there's no predicting what the code
in the GROUP! will change it must invalidate the cache and re-fetch
it after the code has run.

This also adds some debug checking to make sure this circumstance
does not arise again, by making sure each expression evaluation which
could execute arbitrary code does not have a prior frame with a
cached pointer.  That found another case, with variadic execution
doing an execution on a frame somewhere up the stack, but after the
execution was done leaving the f->gotten there where an arbitrary
number of executions might happen before it was used.

To make bugs less likely (and source easier to read), this goes ahead
and ties setting f->gotten to END into the Fetch_Next_In_Frame()
operation.  If for some reason that is wasteful in some cases, that
can be reviewed and a Fetch_Next_In_Frame_Trash_Gotten() could be
added which would only in the debug build trash it...in case the
immediate next operation is to fetch it.  But it doesn't seem worth
it at this point in time.

Includes a small tidying of comments for Push_Frame().