Update chunk stack implementation to reused cell array
The GC in R3-Alpha had no way of knowing how many arguments of a
function had been fulfilled by the evaluator...it could be partway
through or not even started.  So it did a loop, formatting the cells
beforehand...then looping again to actually fill them.

Ren-C had an optimization opportunity, because it was already keeping
a C structure state modeling the progress of a frame.  By letting the
GC have access to this structure, it could know how far in the argument
processing the frame was...and only mark the fulfilled arguments.
This meant the arguments could be filled in one pass.

Trying to preserve this optimization led to a bit of a problem when
the idea of "cell formatting" arose.  This meant that there was a
pattern of bits underlying a cell's header, saying whether it lived
in a heap array or in a stack cell.  If it were left entirely to the
evaluator to format the bits of cells, it would have to choose whether
to format them as heap or stack.

The choice went with was stack, but this led to the unfortunate case
that you couldn't pass in a heap-allocated frame to be filled by the
evaluator loop without corrupting it.  Holding out hope to be able
to preserve the hard-won optimization, some bad hacks were done to
get by in the interim...simulating the evaluator into heap structures
until a solution was made that didn't require preformatting the cells.

This commit implements a hopeful resolution, by changing the nature
of the chunk stack to be implemented with all cells.  This means that
as chunks are subdivided out of a "chunker", the header data is not
some other kind of C structure...but a cell that can be repurposed
as a data-bearing cell on a subsequent push.  This means that the
cell array can be formatted once on chunker allocation, and not walked
again for formatting on each push.

The result is the ability to remove all the Prep_Stack_Cell calls
from %c-eval.c - which is a prerequisite to using it with heap-based
FRAME! structures (e.g. to create specializations from instances, the
way DOES and MAYBE do--currently using hacks).