Boot doesn't depend on DOES specialization, misc cleanup
There was a boot dependency on DOES CATCH [...] which is a little bit
premature, as the feature of doing such specializations is in early
development and has a number of problems.  In order to work on the
feature, it can't be hindering boot.

Has a couple of minor cleanups, including a better way of getting
around a compiler warning using a goto in chunk stack allocation.