Make ELSE not blankify, eliminate EITHER/OPT + EITHER*
The reason that `if true []` is BLANK! is based on a fundamental
necessity.  If NULL were not reserved as the signal for when the
condition was false and the branch did not run, there would be nothing
for ELSE to trigger off of.

However, the rule of necessity does not apply to ELSE.  This means
that if you write:

    x: if condition [...] else [first block]

By *not* blankifying the ELSE result, the nullness can be "telegraphed"
out and lead to an error saying x can't be assigned a null...which is
a good thing.

Yet this is asymmetrical, in the sense that the true branch has no
such protection.  However, the increased usages of ELSE directly mean
this is less of a problem, as decoupling from IF means that expressions
more often look like:

    x: select data item else [first block]

So the value-ness is direct as opposed to coming from a branch.

As a result of this, chaining ELSEs now has meaning:

    x: select data item else [first block] else [<give-up>]

You can look at it as being the ELSE's job to come up with a value
(at the very least a BLANK!) and if it can't do that the chain will
keep on going.

This also implements a change which makes EITHER always return the
exact result of its branches if NULL, instead of blankifying them.  It
creates a small incompatibility with IF...ELSE, because a true branch
of an IF is blankified.

**However**, in the course of this commit the idea of rethinking the
situation so that IF's branches are simply not allowed to evaluate to
null came up.  This would require that procedures return some result
besides NULL, and that idea is now being considered.