IS_VOID() => IS_NULLED()
The change in rules for what had been the "non-ANY-VALUE!" known as
void shifted to being more "null-like", as it was being denoted in the
API with an actual null pointer.  While some of the names were changed,
there had to be a name for the cell holding an indicator of a null
(when it had to be held in a cell).  IS_VOID() was kept around as a
tentative name for that cell state.

However, this made reading code confusing.  Calling it a "nulled cell"
helps avoid confusion between IS_NULL(v) and v == NULL, because the
IS_NULLED(v) name is more suggestive that it is a state of bits in
the cell as opposed to testing the pointer to the cell itself.

This also opens up VOID! for use as an actual datatype, which is now
under consideration as a possible way to sidestep the need for
"blankification".