Make procedures return a VOID! value type
This adds VOID! as the the result given by PROCEDURE calls.  Like a
NULL, it doesn't allow casual usage to assign via SET-WORD!, SET-PATH!,
or a plain SET.  Unlike NULL, it is an ANY-VALUE!, and hence it is
mechanically possible to insert it into blocks.  Also unlike NULL
(which is falsey in conditionals) it is neither true nor false, so:

    if print "Hi" [
        ;-- illegal condition
    ]

Hence as a feature, this brings back the illegality of PRINT in
conditions.  But with the other changes, it also makes it illegal in
ANY and ALL...so printing without output requires ELIDE PRINT (or just
use DUMP instead).