Replace "blankification" with "voidification"
This is a simplifying change which should largely go unnoticed in most
usages.  However it results in the disappearance of a large number of
/OPT refinements and functions with stars in their name (SWITCH*, IF*,
etc.) which had arisen; with hopefully better functionality as a
result too.

ELSE was an experiment in Ren-C that began driving the idea that it
was necessary to be able to tell from the outside of a conditional
expression (like an IF, CASE, SWITCH) whether it took a branch or not.
The concept was to use NULL as a signal for this--since it did not
represent a value.

Yet since NULL was an evaluative result, it was possible to get one
in the branch.  This situation arose:

    if true [
       ;-- nothing (or whatever makes a null result)
    ] else [
       print "this shouldn't run, but would if it got null"
    ]

To work around this, the idea of "BLANK!-ification" came up.  If a
branch ran and returned null, it would be converted into a blank.  You
would lose the distinction, but nulls were not particularly common,
so it seemed a minor issue.

However, NULLs rose to be a major signal--and auto-converting them
to blanks proved to not be always what you wanted.  Trying to deal
with the variations in needs led to the creation of /OPT refinements,
to say that one wanted the branch result "as is" even if it might
produce a NULL and conflate with the output of a non-taken branch.

The blankification came to be applied to every branch...not just the
IF, but the ELSE as well.  In order to avoid casual errors, EITHER
was subjected to the same rules...requiring a special refinement to
ask to pass through the null:

    >> either true [] []
    == _

    >> either false [] []
    == _

    >> either/opt true [] []
    ;-- null

    >> either* false [] [] ;-- * specialization same as /OPT
    ;-- null

This commit pushes back on it...drops those refinements and *-versions,
and eliminates barification, in favor of two principles:

1. If something doesn't have to mutate a result for its usage pattern,
   leave it alone (e.g. EITHER, or an ELSE's branch outcome)

2. If a mutation is needed from a null result, produce a VOID! and not
   a blank.

So basically, this hinges on the idea that someone wanting to get a
NULL out of a branch typically wants to do so from the ELSE branch.
Also, VOID!'s unfriendliness is an asset, because these two are now
fairly similar:

    x: if condition [third [a b]] else [second [z]]
    x: either condition [third [a b]] [second [z]]

Whether condition is true or false, neither will permit the assignment.
While the EITHER can give a true NULL back in the true case, the IF
gives back something spiky enough to draw attention to it.  And in the
falsey branches, the NULL can be given unhindered.

Plus the OPT refinement also considers VOID! along with BLANK! as
a candidate for conversion to NULL.  So it's possible to write:

    opt case [
       ... ["a b c"]
       ... [] ;-- no-op
       ... ["d e f"]
    ]

...and not wind up with a VOID! value that would cause errors in
response to the NO-OP.

Includes native enfixing of the EITHER-TEST-VALUE and EITHER-TEST-NULL
renamed as ALSO and ELSE, so that they can be used earlier in bootstrap
and so their purpose is more clear.  ALSO is in the category of
branchers that need to voidify, in order to telegraph the "success"
signal it branched upon.  Also adds SO as the "Lesser ALSO", which
runs a branch if the left is not null but does not alter the result.