Consolidate singular allocation as Alloc_Singular()
Singular arrays are those which are able to fit a terminated cell's
worth of data into a series node itself, without needing an additional
data allocation.

When first introduced they were used sparingly, but now they are used
pervasively in the system.  The name `Alloc_Singular_Array()` is a bit
redundant, especially as `REBARR *a = Alloc_Singular_Array(), so this
cuts that down to just `Alloc_Singular()`.

Making the name shorter makes it more reasonable to ask all callsites
to include any mask flags they're using.  This makes it clear when a
singular array should have associated file and line information.  Until
now this was done with Alloc_Singular_Array_Core(), but it's a relevant
enough concern to basically all callsites that making it a parameter
to all Alloc_Singular() calls make sense.

(This commit also fixes ENTRAP which was creating a userspace singular
array and not putting file and line information on it, in a good
example of why this change is beneficial).

Includes removal of a few unnecessary DIDs that were noticed--since a
failed SELECT is now logically false and doesn't need them.