More series as managed to begin with, R3_RECYCLE_TORTURE
The basic Make_Array()-style routines start off creating a manually
managed series.  This is so that it isn't necessary to worry about
protecting it from GC if evaluations are run (e.g. perhaps evaluations
are run to fill the cells in the array).  It becomes implicitly managed
when it is inserted into a value, or automatically freed if there is
an error.

This comes at a cost, however.  There is a list of unmanaged series
that the current running action has made, and the list has to be
tended by removing a tracking pointer whenever a series becomes
managed.  Since there's no guarantee on order in which the arrays
were created, theoretically a linear search would have to find it
and take it out.

But series nodes can be created with the NODE_FLAG_MANAGED bit set, and
never put into that tracking list.  This was the only option for
series creation in R3-Alpha--but it's now offered as an extended way
of doing a create--where the managed flag can be passed in with other
SERIES_FLAG_XXX that you want the series to be tagged with--saving on
masking them in as a separate operation.

While not a terribly significant bottleneck, it helps to go ahead and
manage the series at the time of their creation where possible.  It
also attempts to wrangle some of the series and array copying code,
which needs a thorough vetting at some point--as it has hobbled along
in bits here and there since R3-Alpha, and probably needs a generalized
factoring based on passing around a "copy state"/etc.

This also adds an environment switch R3_RECYCLE_TORTURE which can be
used to tell the system to stress test the GC by running a garbage
collect on every evaluator tick...starting from the very first tick.
(If it were a command line option, it wouldn't start running until
after quite a lot of other code had.)