COMPOSE/REDUCE internal API use stack, eliminate /INTO
This commit cleans up the code for doing COMPOSE and REDUCE, by making
them push their results to the stack vs. create an array node.  This
means the caller can choose how to deal with the data, including just
to examine it and drop the stack without making an array at all...
or to append or insert it to an existing array.

It also gets rid of the /INTO protocol, based on arguments made here:

https://forum.rebol.info/t/stopping-the-into-virus/705

The concern is that /INTO muddies the waters of which routines should
have it and which shouldn't, running the risk of becoming spread
throughout the system in interfaces for every function that produces
a series value.  Making allocations and frees cheaper and working
systemically on performance is a wiser investment than seeing that
idea be taken further.  (Its use was minimal, as it was an R3-Alphaism
that did not exist in Rebol2).