Inline series/array allocations, tidy allocation code
The careful bit patterns used to make up a series node are such that
it only takes an assignment to 3 of the 8 platform pointers that they
consist of to make them ready to write to for data that can fit in
the series node.  That's so minimal that it makes a perfect default
state for a series node being made, even if it's going to get a
dynamic expansion immediately after.

This packs the initialization in a very clear way and puts it into the
header files so it can be inlined, hence the optimizer should be able
to pull it together with any initialization of the data afterward.
(e.g. if a singular array is allocated and then has its value cell
initialized, it should only write the header internally once.  Also
any SET_SER_INFO() that is done immediately after should be folded
in with the bit initialization of the info)

It also allows Alloc_Singular() to build on Make_Array_Core() directly,
as the check for (capacity > 1) can be elided by the optimizer when
Make_Array_Core and Make_Series_Node are inlined...and also setting
the array size to 1 can be folded in with setting the info bits.

Includes a bunch of general tidying, including deleting a workaround
for a compiler bug that isn't relevant anymore and getting rid of an
accomodation for calling the old memory checking routines in the
middle of an allocation.