Make REFLECT and OF native, shorten type action dispatcher
R3-Alpha piggy-backed REFLECT on "type action" dispatch.  When 'TYPE
became a reflector as well, that made the type action dispatcher more
complicated because it had to handle NULLs (since TYPE OF () is an
actual null value), so every "type action" dispatch had to check
specificially for SYM_TYPE.

This gets rid of that overhead by just making REFLECT a native.  It
also makes OF a native, so that TYPE OF X is only one native code
function dispatch.  (It previously built a frame and ran REFLECT as
usermode code, since it had to reverse the parameters...making it
substantially slower than when TYPE-OF was its own independent native.)

Includes also a small speedup for integer addition, and some fixes
pertaining to old uses of the meaning of VOID (intending NULL/NULLED).