Make END a CELL_FLAG vs. NODE_FLAG, "less serious" TRASH-ing
Because END was being tested for in bit patterns that were not actually
cells (but Init_Endlike_Header), it was a bit set in the leading byte
of node headers.  This led toa lost opportunity, in the sense that the
SET_END() operation had to use masking to tweak the leading byte while
preserving other cell flags.

To actually do that, the state of whether a cell is a stack-based node
or not would have to be moved into the first byte to take its place.
Also, TRASH-ing a cell wouldn't be able to mess with NODE_FLAG_FREE,
which actually corrects a bug where trashing cells as an intermediate
state was losing the CELL_FLAG_STACK.

This makes the move and the change, which achieves a slight speedup in
the process..as now, SET_END() is a single byte modification that
leaves all the other bytes in the header alone.