Simplify/rename byte-order sensitive macros
The byte-order sensitive macros were initially geared toward a kind of
arbitrary bit-reading purposes, with ideas like reading spans of bits
out that might cross byte boundaries/etc.  The only use of this masking
was in the VAL_PARAM_CLASS() part of headers, because typesets had too
many other flags to really spare a whole byte for that.  But the macros
wound up being more complicated than they needed to be, so this just
asks that specific case to use masking on a byte and do it itself.

Also, the layout was not consistent in terms of byte order across 32
and 64 bit platforms.  This changes that so that the first 4 bytes of
the header will match the first 4 bytes on a 64 bit platform (of the
same endianness).

The byte access macros thus don't need clearing variations, but
evaluate to lvalues of the byte to assign.  This simplifies a lot of
callsites, as one can just say `FIRST_BYTE(v->header) = x` instead of
having separate masking and assignment steps.