"Ghost" nodes, inline Make_Node()/Free_Node()
Anywhere in C that you use a NULL pointer, you have to be sure not to
dereference it.  This means that often times you will see code of
the form:

    if (p != NULL and Get_Property(p, ...)) {...}

Previously Ren-C used something of a trick to avoid null checking by
creating a "dummy" value and using that instead of NULL.  So imagine
a pointer to some kind of thing that would always say it did not have
the property, in the example above.

That trick was the UNBOUND and SPECIFIED pointers, which allocated a
zero length managed array which was only used for such tests.  But a
real array isn't necessarily ideal for this purpose for two reasons:

1. it gets allocated at runtime so that means that its address isn't
a compile-time constant...meaning that `p == UNBOUND` winds up being
slower than `p == NULL` because the address of unbound gets fetched
from a pointer variable.

2. it will casually pass too many tests, including "are you a node"
and "are you an array", which makes it less useful as an opting out
choice.

Since names like "nulled" and "void" and other things are all taken,
this uses the more whimsical name "GHOST node" for a global series-size
object whose address is known at compile time, and whose bits are
fiddled in a way to make it most useful as a null alternative.

Also shuffles things around a little so that Make_Node()/Free_Node()
can be at least partially inlined, to save on pushing and popping
parameters.