PROC/PROCEDURE via `return: <void>`, LEAVE->RETURN
This folds the behavior that had been provided by PROC and PROCEDURE
into FUNC and FUNCTION, by expanding the spec dialect to support a new
tag.  Hence:

    >> data: copy []

    >> emit: function [return: <void> i [integer!]] [
          if i < 1000 [print "not emitting < 1000" | return]
          append data i
       ]

    >> emit 304
    not emitting < 1000

    >> data
    == []

    >> emit 1020
    ;-- no result shown in console, e.g. a VOID!

    >> data
    == [1020]

So the `<void>` annotation on RETURN: not only causes the suppression
of any results which might "fall out" of the bottom of the function,
it also changes RETURN to be arity 0 instead of arity 1.

This is a convenience, as users could get the same effect from:

    >> emit: function [i [integer!]] [
          if i < 1000 [print "not emitting < 1000" | return void]
          append data i
          void
       ]

Earlier Ren-Cs were more prescriptive about the PROCEDURE/FUNCTION
distinction, which was good because it helped function authors think
about whether any results should be "leaking" out of their function
that a caller shouldn't be able to observe.  The EMIT is a good
example, since a caller might notice they were getting the emit buffer
back via the APPEND result...and could try using it, which was likely
not intended by the author of emit.

However, this is now an optional feature that people can choose to
use or not.  By not introducing a new word, it shortens the vocabulary
(no PROC, PROCEDURE, LEAVE) and makes it easier to create new kinds of
function generators without having to give separate names to the
result-returning vs. non-result returning forms (e.g. METHOD would not
need a PROCMETHOD or similar).

The change is largely simplifying, reduces the total amount of code
as well.