Make GHOST polymorphic, change f->varlist to f->reified
The idea of using a GHOST pointer as an alternative to NULL had a bit
of issues with safety (e.g. forgetting that GHOST was being used for
the disengaged state and testing with or assigning null).  It also
was typed, so there would be needs for GHOST_ARRAY or GHOST_CONTEXT,
etc. which is more complicated in usability than NULL which both C
and C++ permit assigning to any type of pointer.

This makes GHOST a void pointer in the C build, and a polymorphic
class which can cast to target node types in the C++ build.  That
gets rid of GHOST_ARRAY and other variations.  It also introduces the
idea of a GHOSTABLE annotation, based on the not_null C++ class in
the C++ core guidelines:

https://visualstudiomagazine.com/articles/2016/06/01/using-the-not_null-template.aspx

This also changes the frame's holding of a "varlist" array to be more
descriptive in pointing to a REBCTX* which is named "reified".