Add METHOD action generator, CONTEXT OF => BINDING OF
This adds a new action generator that is designed to solve two problems
with the use of FUNCTION in objects.  The first problem is that any
assignments via SET-WORD! are gathered as locals instead of assigning
the object members:

    obj: make object! [
        x: 10
        f: function [] [
            x: 10 ;-- would assign a local, since it's gathered
            y: 20 ;-- would also assign a local
        ]
    ]

The new construct is known as METHOD, and it works around this by
quoting the SET-WORD! on its left and reading the binding of it, and
then binding the body into that object:

    obj: make object! [
        x: 10
        m: method [] [
            x: 10 ;-- would assign obj/x, since it's in binding of `m:`
            y: 20 ;-- would assign a local, not in binding of `m:`
        ]
    ]

The second problem this solves is establishing the initial binding
relationship of functions to objects, to be followed by derived binding
through the inheritance graph.  Previously the MAKE OBJECT! call would
walk the variables at creation time and any ACTION! it found that
did not have a binding would have the object pasted into it as the
binding.  This method was rather ad-hoc, whereas using METHOD makes
it explicit.

One mechanical consequence of this commit is that ACTION! generators
like FUNC, FUNCTION, or DOES do not themselves by default have an
"action!-value-binding".  Hence they don't participate in derivation,
and any bindings they have are fixed.  This can be changed by manually
using BIND.

This also changes CONTEXT OF to BINDING OF.  One good reason is because
there can be more than one context associated with things (e.g. the
META OF object is also a context).  Another good reason is that
`context: context of 'x` doesn't seem as meaningful.