Eliminate VAL_SPECIFIC(), fix DEBUG_CHECK_CASTS in older gcc
The previous commit had an issue in g++ 4.8 regarding cast ambiguity
(that wasn't actually ambiguous).  This does some restructuring to
generally make the code clearer, while also handling that problem.

It also caught an issue where a context-based cast was being used on
an UNBOUND...which isn't a context, and then tested against SPECIFIED.
Looking closer at why this was necessary revealed it was going through
a primitive called VAL_SPECIFIC() which was not being meaningfully
used, as it was being used to implement VAL_SPECIFIER() and then the
only other use was in Get_Var_Core(), where it's clearer to model
the usage as a binding.