Fix ASSERT_CELL_WRITABLE macro
It was incorrectly OR'ing flags together for a test when it wanted
to test the flags independently, as opposed to just seeing one or the
other being set.

Includes changes to LINK and MISC calls to take advantage of the fact
that contexts and paramlists are now willing to go through a SER()
cast without needing varlist/paramlist extraction.

Also updates the R3_ALWAYS_MALLOC message to report if PG_Always_Malloc
was set in the build, which is sometimes more convenient than figuring
out how to spawn a debugger from an IDE with an environment variable
set...but easy to accidentally forget and commit.