Eliminate the chunk stack, use pooled memory
When early problems with data stack instability were discovered in
R3-Alpha, a more persistent way of storing memory for arguments of
functions was needed.  At the time there was an option via CLOSURE!
to store data in an array, but this was seen as potentially too
costly due to how slow allocations were--which meant formatting a
valid series, a tracking node, and some data out of a memory pool
on every single function call.

The "chunk stack" was thus implemented as a measure for getting
stable addresses out of a data structure geared especially for the
stacklike nature of function call allocations.  It went along with
some new groundwork in the core of "Do" to be able to maintain a
marker that the GC could access showing how far arguments of a
fulfilling frame had gotten, which could remove the need for
formatting all the cells safely for GC in advance.

With the ability to see the evaluator state of a running function
as a reified FRAME! becoming much more common, it was also common
to want to feed frames in or out of the evaluator.  The existence
of the chunk stack forced every use of the evaluator to copy
from heap frames into the stack...or back out from stack into
heap.  Advanced features like MATCH, specializing DOES, and the
ENCLOSE function generator wound up using hackery to simulate
the evaluator and/or make several unnecessary copies of argument
state.

Series nodes also became much more expressive and versatile, so
this goes ahead and scraps the chunk stack in favor of using
plain allocations from the memory pool.  While a little slower
at the outset, it is a better path for the constructs involved,
and any performance advancements to series allocation will
improve the entire system--as opposed to trying to make the
chunk stack better, which would only help the evaluator.