Don't use singular arrays for paramlists/varlists
The small-series optimization saves on memory use for series that can
fit inside a REBSER node--which is good for things like strings/words
that are small...or for one-item arrays so that `compose [a (b) c]` is
not terribly more expensive than `reduce ['a b 'c]`, so people can
choose which they think fits better.

However, it comes with the cost that dereferencing series has to check
a flag to see if it's a small series or not.  This affects getting
pointers into the data, measuring the length, etc.

On most series it's probably worth it, but the rare case of optimizing
empty parameter lists for functions and empty contexts it isn't.  If
it's known that all contexts and actions use dynamically sized data,
then that can speed up variable indexing and parameter indexing.  So
this goes ahead and does that.

The way it works is to migrate SERIES_INFO_HAS_DYNAMIC into a
SERIES_FLAG, so that it can easily be passed into Make_Series_Core()
and Make_Array_Core().  This way it can cue an override of the small
series optimization...but also be folded out by inlining at the
callsites which don't use it.

Also changes the places where series inaccessibility is checked to be
at the extractions like VAL_CONTEXT() and VAL_ARRAY() instead of on
every raw series operation.