Eliminate Series_Allocation_Unpooled(), header reorganization
This commit's original intended purpose was to remove support for
series whose "wide" was not a power of 2.  The original R3-Alpha code
had no series that didn't follow this rule, but the addition of the
FFI led to allocations of memory for structs (not REBYTE, REBUNI,
or REBVAL) which had arbitrary sizes.  The code worked, but the fact
that sometimes series were allocated to sizes rounded to powers of 2
and allocated via malloc() (vs. in a known-sized pool) meant that the
"rounded down" fields inside the series node (rest, bias, len) could
not be used to recreate the actual count of bytes made for the series.

The problem was noted when trying to add a check that on exit, Rebol's
notion of how much memory allocation was outstanding balanced to zero.
If it didn't, then the decisions made on when to GC or not could
arbitrarily drift until the number representing the outstanding memory
was essentially meaningless.  So it was important to get that right.

Because it would throw the size of series nodes off to store their
"actual" allocation in bytes, the workaround was to just use a bit
to record the fact that a series *had* been allocated to a power of 2
size.  This meant that the same rounding process that produced the
allocation for the unevenly sized struct element could be run again
to generate the number.

In doing performance analysis of work Ren-C is doing that R3-Alpha
isn't, this showed up on the radar as a performance cost.  The feature
it was intended to support--non-power-of-2-size series--appears to no
longer be in use.  This simply drops support for cases where a series
size can't be determined from the REBSER.content.dynamic alone.

Deleting the routine and putting SER_TOTAL() into the headers required
some rearrangement, however.  This rearrangement also pushed the
definition of REBSYM out of the internal API, which paves a path
forward for "type actions" to take arbitrary words vs. just the ones
that appear in %words.r or in type action "verbs".