Frame reuse: Fixes MATCH bug, optimize ENCLOSE, speeds APPLY
This is a first cut at addressing issues regarding frame reuse in the
evaluator.  By actually building a memory layout which can be
compatible with ANY-CONTEXT! in the evaluator,

(Note: This is somewhat similar in spirit to what R3-Alpha was doing
with CLOSURE! by evaluating into a heap object.  However, there are
significant details regarding using GC management bit status to do
reuse of these objects, as well as to be able to not pre-initialize
bit patterns in the array due to frames, along with finessing aspects
of how frame "phases" and "bindings" are kept in sync.)

The main thrust of the change is to allow frames to be "stolen" back
and forth in the evaluation process.  This means that something like
MATCH can do this:

     >> match parse "aaa" [some "a"]
     == "aaa"

What happens under the hood is that MATCH asks PARSE to evaluate its
arguments into a FRAME!, but not execute it.  Then it grabs the first
argument out of the frame, before running it and determining if its
result is truthy or falsey.  If truthy, it returns that first
argument it captured.

Previously such mechanics were not easily supported by the evaluator.
Now with a little bit of fiddling, the evaluator can be told to run
but then stop before actually executing or dropping the frame...and
functions like MATCH can then "steal" that directly...get the values,
and then turn around and run the frame again.

Because stealing in this fashion was not available before, the process
of an ENCLOSE had to make several copies of argument lists.  Now it
uses the same memory throughout--one allocation to gather interface
parameters, which is then reclaimed to become the FRAME! passed to
the user function, and if the user chooses to DO that FRAME! (with
possible modifications) then that storage is used again as well.  This
cements the protocol that DO of a FRAME! will make it inaccessible,
and it's the user's responsibility to COPY and pay that cost to have
the ability to use the frame again after the DO.