Fix infinite loop in charset generation
The loop that set bits in `charset` was using a REBUNI value to do
the loop, and comparing against that value while sweeping through the
bits to set.  But if the last bit was the maximum REBUNI value, then
it would overflow back to 0 instead of realizing it was the limit.
Hence this was an infinite loop:

    charset [#"^(FFFE)" - #"^(FFFF)"]

While there are changes pending which would make REBUNI use more than
a 2-byte representation, this just goes ahead and tweaks the loop so
that this case works correctly for now...by using post-increment after
the loop thus testing the condition before potential overflow is seen.