Small efficiency trick: no conditional for f->special assign
In the "every bit helps" category, this avoids one conditional
expression in every action call related to whether or not the action
has specialization data.  Rather than store a null pointer in the
"exemplar" slot for this case, it stores a pointer to the "facade"...
which basically means that whatever the [1] element of that array is
would be the correct thing to assign to f->special for the purposes
of Do_Core().

As this is slightly obscuring, it changes the field name stored in the
series node from 'exemplar' to 'specialty', and the type from REBCTX
to REBARR.