%tmp-funcs.h => %tmp-internals.h, break out %sys-rebctx.h
Ren-C required large-scale reorganizations in order to use inline
functions instead of macros.  Because while a macro has no dependencies
on the text it substitutes, an inline function in C is otherwise like
a regular function...you can't use anything that isn't defined yet,
with the only way to bend that rule to be talking about something by
pointer that has been "forward-declared".

As a consequence, the header files had to react by breaking up into
dependent groups.  One thing that happened was that because REBSER
nodes included a REBVAL cell by value, REBVALs had to be defined before
REBSER...but many value accessors needed to do things like react to
the series flags in the contained payload.  This led to a split into
%sys-rebval.h, %sys-rebser.h, %sys-rebser.h and %sys-value.h...where
rebser could depend on rebval but then value macros could depend on
series macros.

Other types wound up less formally involved in this, but this takes
a further step in making the separation clearer with a separate file
for %sys-rebctx.h from %sys-context.h.  It improves the comments
somewhat, and also makes it clear that a key differentiating factor
in deciding what-goes-where is whether a definition depends on having
the internal API functions defined or not.  It also renames the
internal API file %tmp-internals.h instead of %tmp-funcs.h, as it is
less vague.